containers: replace TransferId with ConsignmentId
