+++ src/containers/consignment.rs
@@ -20,10 +20,17 @@
 // limitations under the License.
 
 use std::collections::BTreeMap;
-use std::iter;
+use std::fmt::{Display, Formatter};
+use std::str::FromStr;
+use std::{fmt, iter};
 
-use amplify::confinement::{LargeOrdSet, MediumBlob, SmallOrdMap, TinyOrdMap, TinyOrdSet};
+use amplify::confinement::{
+    LargeOrdSet, MediumBlob, SmallOrdMap, SmallOrdSet, TinyOrdMap, TinyOrdSet,
+};
+use amplify::{ByteArray, Bytes32};
+use baid58::{Baid58ParseError, Chunking, FromBaid58, ToBaid58, CHUNKING_32};
 use bp::Tx;
+use commit_verify::{CommitEncode, CommitEngine, CommitmentId, DigestExt, Sha256};
 use rgb::validation::{self};
 use rgb::{
     AnchoredBundle, AssetTag, AssignmentType, AttachId, BundleId, ContractHistory, ContractId,
@@ -42,6 +49,62 @@ use crate::LIB_NAME_RGB_STD;
 
 pub type Transfer = Consignment<true>;
 pub type Contract = Consignment<false>;
+
+/// Interface identifier.
+///
+/// Interface identifier commits to all the interface data.
+#[derive(Wrapper, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash, Debug, From)]
+#[wrapper(Deref, BorrowSlice, Hex, Index, RangeOps)]
+#[derive(StrictType, StrictDumb, StrictEncode, StrictDecode)]
+#[strict_type(lib = LIB_NAME_RGB_STD)]
+#[cfg_attr(
+    feature = "serde",
+    derive(Serialize, Deserialize),
+    serde(crate = "serde_crate", transparent)
+)]
+pub struct ConsignmentId(
+    #[from]
+    #[from([u8; 32])]
+    Bytes32,
+);
+
+impl From<Sha256> for ConsignmentId {
+    fn from(hasher: Sha256) -> Self { hasher.finish().into() }
+}
+
+impl CommitmentId for ConsignmentId {
+    const TAG: &'static str = "urn:lnp-bp:rgb:consignment#2024-03-11";
+}
+
+impl ToBaid58<32> for ConsignmentId {
+    const HRI: &'static str = "con";
+    const CHUNKING: Option<Chunking> = CHUNKING_32;
+    fn to_baid58_payload(&self) -> [u8; 32] { self.to_byte_array() }
+    fn to_baid58_string(&self) -> String { self.to_string() }
+}
+impl FromBaid58<32> for ConsignmentId {}
+impl Display for ConsignmentId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        if !f.alternate() {
+            f.write_str("urn:lnp-bp:consignment:")?;
+        }
+        if f.sign_minus() {
+            write!(f, "{:.2}", self.to_baid58())
+        } else {
+            write!(f, "{:#.2}", self.to_baid58())
+        }
+    }
+}
+impl FromStr for ConsignmentId {
+    type Err = Baid58ParseError;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Self::from_baid58_maybe_chunked_str(s.trim_start_matches("urn:lnp-bp:"), ':', '#')
+    }
+}
+impl ConsignmentId {
+    pub const fn from_array(id: [u8; 32]) -> Self { Self(Bytes32::from_array(id)) }
+    pub fn to_mnemonic(&self) -> String { self.to_baid58().mnemonic() }
+}
 
 /// Consignment represents contract-specific data, always starting with genesis,
 /// which must be valid under client-side-validation rules (i.e. internally
@@ -113,6 +176,34 @@ pub struct Consignment<const TYPE: bool> {
 
 impl<const TYPE: bool> StrictSerialize for Consignment<TYPE> {}
 impl<const TYPE: bool> StrictDeserialize for Consignment<TYPE> {}
+
+impl<const TYPE: bool> CommitEncode for Consignment<TYPE> {
+    type CommitmentId = ConsignmentId;
+
+    fn commit_encode(&self, e: &mut CommitEngine) {
+        e.commit_to_serialized(&self.version);
+        e.commit_to_serialized(&self.transfer);
+
+        e.commit_to_serialized(&self.contract_id());
+        e.commit_to_serialized(&self.genesis.disclose_hash());
+        e.commit_to_set(&TinyOrdSet::from_iter_unsafe(
+            self.ifaces.values().map(|pair| pair.iimpl.impl_id()),
+        ));
+
+        e.commit_to_set(&LargeOrdSet::from_iter_unsafe(
+            self.bundles.iter().map(|ab| ab.bundle.disclose_hash()),
+        ));
+        e.commit_to_set(&LargeOrdSet::from_iter_unsafe(
+            self.extensions.iter().map(Extension::disclose_hash),
+        ));
+        e.commit_to_set(&SmallOrdSet::from_iter_unsafe(self.terminals_disclose()));
+
+        e.commit_to_set(&SmallOrdSet::from_iter_unsafe(self.attachments.keys().copied()));
+        e.commit_to_set(&self.supplements);
+        e.commit_to_map(&self.asset_tags);
+        e.commit_to_map(&self.signatures);
+    }
+}
 
 impl<const TYPE: bool> Consignment<TYPE> {
     /// # Panics
