+++ src/containers/mod.rs
@@ -31,7 +31,6 @@
 mod consignment;
 mod disclosure;
 mod bindle;
-mod transfer;
 mod seal;
 mod util;
 mod validate;
@@ -41,7 +40,7 @@ mod indexed;
 
 pub use bindle::{Bindle, BindleContent, BindleParseError, LoadError, UniversalBindle};
 pub use certs::{Cert, ContentId, ContentSigs, Identity};
-pub use consignment::{Consignment, Contract, Transfer};
+pub use consignment::{Consignment, ConsignmentId, Contract, Transfer};
 pub use disclosure::Disclosure;
 pub use indexed::IndexedConsignment;
 pub use partials::{Batch, CloseMethodSet, Fascia, TransitionInfo};
